Create a game provider
Create a provider and make it available to the game page only.
